# openapi.yaml
openapi: 3.0.3
info:
  title: API Validation Prototype
  version: 0.0.1
paths:
  /ping:
    get:
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                type: string

  /time:
    get:
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  ts:
                    type: string

  /echo:
    post:
      requestBody:
        content:
          application/json:
            schema:
              type: object
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - type: object
                    properties:
                      ts:
                        type: string
                  - type: object
                    properties:
                      body:
                        type: string
                      headers:
                        type: object

  /valid:
    post:
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                mandatory:
                  type: string
                optional:
                  type: string
              required:
                - mandatory
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                type: object
        400:
          description: Failed
          content:
            application/json:
              schema:
                type: object

components:
  schemas:

    Types_Foo:
      type: string
      enum:
        - a
        - s
        - d
        - f

    Types_Bar:
      type: string
      enum:
        - q
        - w
        - e
        - r
        - t

    Types:
      anyOf:
        - $ref: "#/components/schemas/Types_Bar"
        - $ref: "#/components/schemas/Types_Foo"